# Optimize and cache the signature computation for update.updateFromExternalObject