--- a/src/nti/externalization/externalization/externalizer.py
+++ b/src/nti/externalization/externalization/externalizer.py
@@ -252,7 +252,7 @@ def _to_external_object_state(obj, state, top_level=False):
         elif result is not _marker:
             return result
         else:
-            logger.warn("Recursive call to object %s.", obj)
+            logger.warning("Recursive call to object %s.", obj)
             result = internal_to_standard_external_dictionary(obj,
                                                               decorate=False)