--- a/src/nti/externalization/externalization/tests/test_externalizer.py
+++ b/src/nti/externalization/externalization/tests/test_externalizer.py
@@ -40,7 +40,7 @@ class TestFunctions(unittest.TestCase):
         class Proxy(ProxyBase):
 
             def toExternalObject(self, **kwargs):
-                return {'a': 42}
+                raise NotImplementedError
 
         self.assertFalse(_obj_has_usable_externalObject(Proxy(object())))