--- a/src/nti/externalization/tests/benchmarks/bm_simple_registered_class.py
+++ b/src/nti/externalization/tests/benchmarks/bm_simple_registered_class.py
@@ -17,6 +17,22 @@ from nti.externalization.internalization import update_from_external_object
 import nti.externalization.tests.benchmarks
 from nti.externalization.tests.benchmarks.objects import SimplestPossibleObject
 
+# pylint:disable=arguments-differ
+
+class NoArgs(SimplestPossibleObject):
+
+    def updateFromExternalObject(self, external_object):
+        self.__dict__.update(external_object)
+
+class ContextArg(SimplestPossibleObject):
+
+    def updateFromExternalObject(self, external_object, context=None):
+        self.__dict__.update(external_object)
+
+class DSArg(SimplestPossibleObject):
+
+    def updateFromExternalObject(self, external_object, dataserver=None):
+        self.__dict__.update(external_object)
 
 def main(runner=None):
 
@@ -38,6 +54,27 @@ def main(runner=None):
         update_from_external_object(obj, ext)
 
     runner.bench_func(__name__ + ": fromExternalObject", from_)
+
+
+    def no_args():
+        obj = NoArgs()
+        update_from_external_object(obj, ext)
+
+    def context_arg():
+        obj = ContextArg()
+        update_from_external_object(obj, ext)
+
+    def ds_arg():
+        obj = DSArg()
+        update_from_external_object(obj, ext)
+
+
+    runner.bench_func(__name__ + ": fromExternalObject (no args)", no_args)
+
+    runner.bench_func(__name__ + ": fromExternalObject (context arg)", context_arg)
+
+    runner.bench_func(__name__ + ": fromExternalObject (ds arg)", ds_arg)
+