This lets us be 8-27% faster.

Also officially deprecate with a warning some of the older signatures. This addresses (but doesn't quite fix) #30 

| Benchmark | 27_src_all_no_cache | 27_src_all_caches_3          |
|---------------------------------------------------------------------------------------------------|---------------------|------------------------------|
| .bm_simple_iface: find factory                                | 4.01 us             | 3.96 us: 1.01x faster (-1%)  |
| .bm_simple_iface: fromExternalObject                          | 70.4 us             | 65.0 us: 1.08x faster (-8%)  |
| .bm_simple_registered_class: fromExternalObject               | 13.6 us             | 10.9 us: 1.25x faster (-20%) |
| .bm_simple_registered_class: fromExternalObject (no args)     | 21.2 us             | 16.2 us: 1.31x faster (-24%) |
| .bm_simple_registered_class: fromExternalObject (context arg) | 21.3 us             | 16.2 us: 1.32x faster (-24%) |
| .bm_simple_registered_class: fromExternalObject (ds arg)      | 22.1 us             | 16.1 us: 1.37x faster (-27%) |

Not significant (6): Construct Singleton; Construct non-Singleton; Construct Singleton with args; Construct non-Singleton with args; nti.externalization.tests.benchmarks.bm_simple_iface: toExternalObject; nti.externalization.tests.benchmarks.bm_simple_registered_class: toExternalObject