@jamadden 2018-07-10
I'll expect to merge this in the morning if I receive no feedback.